# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend Challenge",
        "version": "1.0.0",
        "title": "Turing/E-commerce",
        "contact": {
            "email": "codeninja0x01@gmail.com"
        }
    },
    "basePath": "/api",
...
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "name": "USER-KEY",
            "in": "header"
        }
    },
    "host": "localhost:3000",
    "paths": {
...
        "/attributes/product/{productId}": {
            "get": {
                "summary": "Get Attribute product.",
                "description": "Return a list of attribute product.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "productId",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AttributeProduct"
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid user input."
                    }
                },
                "tags": [
                    "attribute"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
...
        "/categories/product/{productId}": {
            "get": {
                "summary": "Get category of product.",
                "description": "Return category by productId.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "productId",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Category"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "category"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
...
        "/products": {
            "get": {
                "summary": "Get list of products.",
                "description": "Return products.",
                "parameters": [
                    {
                        "$ref": "#/parameters/limitParam"
                    },
                    {
                        "$ref": "#/parameters/offsetParam"
                    },
                    {
                        "$ref": "#/parameters/orderParam"
                    },
                    {
                        "$ref": "#/parameters/orderByParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "$ref": "#/definitions/ProductPagination"
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "product"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/products/search": {
            "get": {
                "summary": "Search products.",
                "description": "Return products.",
                "parameters": [
                    {
                        "$ref": "#/parameters/searchParam"
                    },
                    {
                        "$ref": "#/parameters/limitParam"
                    },
                    {
                        "$ref": "#/parameters/offsetParam"
                    },
                    {
                        "$ref": "#/parameters/descLengthParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "$ref": "#/definitions/ProductPagination"
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "product"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/products/{id}": {
            "get": {
                "summary": "Get product.",
                "description": "Return product by id.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "id",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "product"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/products/category/{categoryId}": {
            "get": {
                "summary": "Get category products.",
                "description": "Return products by categoryId.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "categoryId",
                        "required": true,
                        "in": "path"
                    },
                    {
                        "$ref": "#/parameters/limitParam"
                    },
                    {
                        "$ref": "#/parameters/offsetParam"
                    },
                    {
                        "$ref": "#/parameters/orderParam"
                    },
                    {
                        "$ref": "#/parameters/orderByParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "product"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/products/department/{departmentId}": {
            "get": {
                "summary": "Get department products.",
                "description": "Return products by departmentId.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "departmentId",
                        "required": true,
                        "in": "path"
                    },
                    {
                        "type": "number",
                        "name": "offset",
                        "required": false,
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "limit",
                        "required": false,
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "orderBy",
                        "required": false,
                        "in": "query",
                        "enum": [
                            "name",
                            "id"
                        ],
                        "default": "name"
                    },
                    {
                        "type": "string",
                        "name": "order",
                        "required": false,
                        "in": "query",
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "default": "ASC"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "product"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/products/{id}/reviews": {
            "get": {
                "summary": "Get product reviews.",
                "description": "Return product reviews.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "id",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ProductReview"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "product"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            },
            "post": {
                "summary": "Add product review.",
                "description": "Create product reviews.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "id",
                        "required": true,
                        "in": "path"
                    },
                    {
                        "name": "review",
                        "required": false,
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "rating",
                        "required": true,
                        "in": "formData",
                        "type": "number"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ProductReview"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    },
                    "401": {
                        "description": "Unauthorized User."
                    }
                },
                "tags": [
                    "product"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
        "/products/{id}/location": {
            "get": {
                "summary": "Get product location.",
                "description": "Return product location.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "id",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ProductLocation"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "product"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
...
    },
    "parameters": {
        "limitParam": {
            "name": "limit",
            "in": "query",
            "description": "Limits the number of returned results",
            "required": false,
            "type": "integer",
            "default": 20
        },
        "offsetParam": {
            "name": "offset",
            "in": "query",
            "description": "Offset from which start returned results",
            "required": false,
            "type": "integer",
            "default": 1
        },
        "orderParam": {
            "name": "order",
            "in": "query",
            "description": "Order ASC or Desc the results",
            "required": false,
            "type": "string",
            "enum": [
                "ASC",
                "DESC"
            ]
        },
        "orderByParam": {
            "name": "orderBy",
            "in": "query",
            "description": "OrderBy name or id the results",
            "required": false,
            "type": "string",
            "enum": [
                "id",
                "name"
            ]
        },
        "searchParam": {
            "name": "search",
            "in": "query",
            "description": "String to search in the database",
            "required": true,
            "type": "string"
        },
        "descLengthParam": {
            "name": "descLength",
            "in": "query",
            "description": "Description text length",
            "required": false,
            "type": "string",
            "default": 200
        }
    },
    "definitions": {
...
        "AttributeProduct": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            },
            "required": [
                "id",
                "name",
                "value"
            ]
        },
...
        "Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            },
            "required": [
                "id",
                "name",
                "description"
            ]
        },
...
        "Product": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "discountedPrice": {
                    "type": "number"
                },
                "image1": {
                    "type": "string"
                },
                "image2": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "display": {
                    "type": "boolean"
                }
            },
            "required": [
                "id",
                "name",
                "description",
                "price",
                "discountedPrice",
                "image1",
                "image2",
                "thumbnail",
                "display"
            ]
        },
        "ProductPagination": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "number"
                },
                "currentPage": {
                    "type": "number"
                },
                "perPage": {
                    "type": "number"
                },
                "row": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Product"
                    }
                }
            },
            "required": [
                "total",
                "currentPage",
                "perPage",
                "row"
            ]
        },
        "ProductReview": {
            "type": "object",
            "properties": {
                "review": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "userName": {
                    "type": "string"
                },
                "createdOn": {
                    "type": "string",
                    "format": "date-time"
                }
            },
            "required": [
                "review",
                "rating",
                "userName",
                "createdOn"
            ]
        },
        "ProductLocation": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "type": "number"
                },
                "categoryName": {
                    "type": "string"
                },
                "departmentId": {
                    "type": "number"
                },
                "departmentName": {
                    "type": "string"
                }
            },
            "required": [
                "categoryId",
                "categoryName",
                "departmentId",
                "departmentName"
            ]
        },
...
    }
}